+++ integration-tests/spark_int_test/Dockerfile.int-spark
@@ -27,16 +27,17 @@ RUN apt-get update -y -qq && apt-get install -y -qq netcat
 # instrumentation:
 RUN pip install -U pytest sh pyspark tox 'pip>=20,<20.3'
 
+ARG AIRFLOW_EXTRA=airflow
 # install dbnd packages:
 ADD ./dist/ /dist/
 RUN pip install -r /dist/dbnd.requirements.txt && \
     pip install -r /dist/dbnd-spark.requirements.txt && \
     pip install -r /dist/dbnd-test-scenarios.requirements.txt && \
     pip install -r /dist/dbnd-airflow.requirements.txt && \
-    pip install -r /dist/dbnd-airflow[airflow].requirements.txt && \
+    pip install -r /dist/dbnd-airflow[$AIRFLOW_EXTRA].requirements.txt && \
     pip install -r /dist/dbnd-airflow[tests].requirements.txt && \
     pip install -r /dist/examples.requirements.txt
-RUN pip install dbnd_airflow[airflow] dbnd_examples databand[spark] dbnd_test_scenarios --find-links /dist/ --no-index
+RUN pip install dbnd_airflow[$AIRFLOW_EXTRA] dbnd_examples databand[spark] dbnd_test_scenarios --find-links /dist/ --no-index
 
 ENV DBND_EXAMPLES_PATH=/app/examples
 WORKDIR /app/
