--- modules/dbnd-airflow/src/dbnd_airflow/__main__.py
@@ -1,5 +1,0 @@
-from dbnd_airflow.dbnd_airflow_main import main
-
-
-if __name__ == "__main__":
-    main()
