+++ modules/dbnd-airflow/tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = {py27,py36,py37}-{airflow_1_10_7,airflow_1_10_9,airflow_1_10_10, airflow_1_10_12},manifest
+envlist = {py27,py36,py37}-{airflow_1_10_7,airflow_1_10_9,airflow_1_10_10, airflow_1_10_12, airflow_1_10_15},manifest
 toxworkdir = {toxinidir}/../../.tox/dbnd-airflow
 
 [dbnd]
@@ -40,6 +40,7 @@ deps =
 	airflow_1_10_9: apache-airflow==1.10.9
 	airflow_1_10_10: apache-airflow==1.10.10
     airflow_1_10_12: apache-airflow==1.10.12
+    airflow_1_10_15: apache-airflow==1.10.15
 
 	airflow_1_10_1: marshmallow-sqlalchemy==0.18.0
 	airflow_1_10_1: werkzeug<1.0.0
