+++ modules/dbnd/src/dbnd/testing/helpers.py
@@ -75,7 +75,7 @@ def run_test_notebook(notebook):
 
 def run_subprocess__airflow(args, retcode=255, **kwargs):
     return run_dbnd_subprocess(
-        args=["dbnd-airflow"] + args,
+        args=["airflow"] + args,
         cwd=kwargs.pop("cwd", os.environ[ENV_DBND_HOME]),
         retcode=retcode,
         **kwargs
