+++ plugins/dbnd-airflow-versioned-dag/tox.ini
@@ -35,7 +35,7 @@ commands =
 	rm -f {env:DBND_TEST_DB}
 	coverage erase
 # we don't need databand db, only airflow
-	dbnd-airflow initdb
+	airflow initdb
 	pytest --cov dbnd_airflow_versioned_dag --junit-xml build/junit-{envname}.xml {[dbnd]tests} {posargs}
 	coverage report
