+++ plugins/dbnd-aws/tox.ini
@@ -49,7 +49,7 @@ filterwarnings =
 commands =
 	rm -f {env:AIRFLOW_TEST_DB}
 	# we don't need databand db, only airflow
-	dbnd-airflow initdb
+	airflow initdb
 
 	coverage erase
 	pytest --cov dbnd_aws --run-aws  --junit-xml build/junit-{envname}.xml {[dbnd]tests} {posargs}
@@ -61,7 +61,7 @@ commands =
     rm -f {env:AIRFLOW_TEST_DB}
 	coverage erase
 	# we don't need databand db, only airflow
-	dbnd-airflow initdb
+	airflow initdb
 	pytest --cov dbnd_aws --run-aws --run-emr --junit-xml build/junit-{envname}.xml {[dbnd]tests_emr} {posargs}
 	coverage report
