+++ plugins/dbnd-databricks/tox.ini
@@ -42,7 +42,7 @@ filterwarnings =
 
 commands =
 	rm -f {env:AIRFLOW_TEST_DB}
-	dbnd-airflow initdb
+	airflow initdb
 	pytest -m "not databricks_integration" --cov dbnd_databricks --junit-xml build/junit-{envname}.xml {[dbnd]tests} {posargs}
 	coverage report
