+++ plugins/dbnd-luigi/setup.py
@@ -14,7 +14,7 @@ version = config["metadata"]["version"]
 setuptools.setup(
     name="dbnd-luigi",
     package_dir={"": "src"},
-    install_requires=["dbnd==" + version, "luigi"],
+    install_requires=["dbnd==" + version, "luigi<=3.0.2"],
     extras_require={"test": ["docutils"]},  # Indirect luigi dependency
     entry_points={
         "console_scripts": ["dbnd-luigi = dbnd_luigi.luigi_tracking:dbnd_luigi_run",],
