+++ plugins/dbnd-spark/tox.ini
@@ -38,7 +38,7 @@ filterwarnings =
 
 commands =
 	rm -f {env:AIRFLOW_TEST_DB}
-	dbnd-airflow initdb
+	airflow initdb
     pytest --cov dbnd_spark --junit-xml build/junit-{envname}.xml {[dbnd]tests} {posargs}
 ;	coverage run --parallel -m pytest --junit-xml build/junit-{envname}.xml {[dbnd]tests} {posargs}
 ;    coverage run --parallel -m pytest --run-spark --junit-xml build/junit-{envname}.xml {[dbnd]tests} {posargs}
