+++ plugins/dbnd-tensorflow/tox.ini
@@ -40,7 +40,7 @@ filterwarnings =
 
 commands =
 	rm -f {env:DBND_TEST_DB}
-	dbnd-airflow initdb
+	airflow initdb
 	coverage run --parallel -m pytest --junit-xml build/junit-{envname}.xml {[dbnd]tests} {posargs}
